Initial commit to change package structure towards v1.1 direction
