+++ src/main/java/owlmoney/logic/command/Command.java
@@ -1,0 +1,32 @@
+package owlmoney.logic.command;
+
+import owlmoney.logic.exception.DukeException;
+import owlmoney.storage.Storage;
+import owlmoney.model.task.TaskList;
+import owlmoney.ui.Ui;
+
+/**
+ * Abstract class representing individual commands.
+ */
+public abstract class Command {
+    boolean exit = false;
+
+    /**
+     * Determines whether this command is exiting, i.e. whether the program should exit
+     * after executing this command.
+     *
+     * @return Whether this command is exiting.
+     */
+    public boolean isExit() {
+        return exit;
+    }
+
+    /**
+     * Executes this command on the given task list and user interface.
+     *
+     * @param tasks The task list.
+     * @param ui The user interface displaying events on the task list.
+     * @param storage The place where tasks will be stored.
+     */
+    public abstract void execute(TaskList tasks, Ui ui, Storage storage) throws DukeException;
+}
